Move  setup funcs to their own file
